basic comix connections
This script creates three images: "image1.jpg", "link.jpg", and "image2.jpg". The first two images are set to be scrollable using the "overflow: scroll;" CSS property.
 The third image, "link.jpg", is also set to be scrollable and has an onclick event that uses JavaScript to scroll the second image, "img2", into view when clicked.

You can change the source of the images, the width and height of the images and the onclick event to connect different images.